requirements.txt
fastapi
uvicorn
openai
python-dotenv
google-generativeai>=0.3.2
websockets
pydantic
